pyproject.toml
[tool.poetry]
name = "plsarchiver"
version = "0.1.0"
description = ""
authors = ["Pierre-Yves Gillier <github@pygillier.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
spotify-client = "==1.9.1"
click = "==8.1.3"
python-dotenv = "==1.0.0"
spotipy = "==2.23.0"
flask = "==2.3.2"
flask-session2 = "==1.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
